feat: Refactor step definitions to inherit from BaseStepDefinitions and streamline flag handling
